gradio_cli_app/pyproject.toml
[project]
name = "gradio-cli-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "certifi>=2025.11.12",
    "google-generativeai>=0.8.5",
    "gradio>=6.0.1",
    "groq>=0.37.0",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
]
